server side auth complete

+++ src/server/db/queries/blogs.ts
@@ -36,7 +36,7 @@ GROUP BY b.id
 	);
 const create = (newBlog: IBaseBlogs) => Query("INSERT INTO Blogs SET ?", [newBlog]);
 const update = (updatedBlog: IBaseBlogs, id: number) => Query("UPDATE Blogs SET ? WHERE id=?", [updatedBlog, id]);
-const destroy = (id: number) => Query("DELETE FROM Blogs WHERE id=?", [id]);
+const destroy = (id: number, author_id: number) => Query("DELETE FROM Blogs WHERE id=? and author_id =?", [id, author_id]);
 
 export default {
 	getAll,
